+++ httpd/service/echo.cpp
@@ -6,7 +6,7 @@ using std::placeholders::_3;
 Echo::Echo(shared_ptr<Server> server_) : server(server_) {
     server->onMessage = bind(&Echo::onMessage, this, _1, _2, _3);
 }
-void Echo::onMessage(char *buf, uint64_t size,
-                     function<ssize_t(char *buf, uint64_t size)> send) {
+void Echo::onMessage(char *buf, ssize_t size,
+                     function<ssize_t(char *buf, ssize_t size)> send) {
     send(buf, size);
 }
